+++ tests/wikitext/test_external_links.py
@@ -213,3 +213,14 @@ def test_external_link_should_not_span_over_tags():
 def test_external_link_in_pf_in_tl():  # 110
     (els,) = parse('{{text|<ref>[https://a.b a]</ref>}}').external_links
     assert els.string == '[https://a.b a]'
+
+
+def test_ext_link_overwriting_template():  # 74
+    p = parse("{{test}}")
+    t = p.templates[0]
+    # It is could be considered unintuitive for p.templates to still return
+    # [Template('[https://link]')]. As it stands, the user who is overwriting
+    # the template should keep track of such changes and skip edited templates
+    # that have become invalid.
+    t.string = "[https://link]"
+    assert p.external_links[0].string == "[https://link]"